feat(AED): Primer test de JPA hecho

+++ AED/JPA/Instituto/src/main/java/es/iespuertodelacruz/saul/entities/Alumno.java
@@ -24,7 +24,7 @@ public class Alumno implements Serializable {
 
 	private String apellidos;
 
-	private BigInteger fechanacimiento;
+	private Long fechanacimiento;
 
 	private String nombre;
 
@@ -51,11 +51,11 @@ public class Alumno implements Serializable {
 		this.apellidos = apellidos;
 	}
 
-	public BigInteger getFechanacimiento() {
+	public Long getFechanacimiento() {
 		return this.fechanacimiento;
 	}
 
-	public void setFechanacimiento(BigInteger fechanacimiento) {
+	public void setFechanacimiento(Long fechanacimiento) {
 		this.fechanacimiento = fechanacimiento;
 	}
